backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
ollama==0.3.3
arxiv==2.1.3
semanticscholar==0.8.4
apscheduler==3.10.4
gunicorn==22.0.0
python-dateutil==2.9.0
scholarly==1.7.0
feedparser==6.0.11
beautifulsoup4==4.12.2  # Added for scholarly parsing
lxml==4.9.4  # Added for HTML parsing
tenacity==8.3.0  # Added for retries
langchain-community
tenacity
tenacity==8.3.0
requests==2.32.3
feedparser==6.0.11
scholarly==1.7.0
tenacity==8.3.0
arxiv==2.1.3
